third_party/blink/web_tests/platform/generic/shadow-dom/css-cascade-slot-distributed-expected.txt

Checking if styles in the nested shadow roots apply properly to distributed elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(span).color is "rgb(0, 128, 0)"
PASS getComputedStyle(span).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

There should be no red
